refactor(deeplink): making the deeplink package public
