+++ telegram/message/peer/resolve.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-faster/errors"
 
 	"github.com/gotd/td/ascii"
-	"github.com/gotd/td/telegram/internal/deeplink"
+	"github.com/gotd/td/telegram/deeplink"
 	"github.com/gotd/td/tg"
 )
